+++ lualib/sproto.lua
@@ -13,14 +13,23 @@ function sproto_mt:__gc()
 	core.deleteproto(self.__cobj)
 end
 
-function sproto.new(bin,sz,nogc)
-	local cobj = assert(core.newproto(bin,sz))
+function sproto.new(bin)
+	local cobj = assert(core.newproto(bin))
 	local self = {
 		__cobj = cobj,
 		__tcache = setmetatable( {} , weak_mt ),
 		__pcache = setmetatable( {} , weak_mt ),
 	}
-	return setmetatable(self, nogc and sproto_nogc or sproto_mt)
+	return setmetatable(self, sproto_mt)
+end
+
+function sproto.sharenew(cobj)
+	local self = {
+		__cobj = cobj,
+		__tcache = setmetatable( {} , weak_mt ),
+		__pcache = setmetatable( {} , weak_mt ),
+	}
+	return setmetatable(self, sproto_nogc)
 end
 
 function sproto.parse(ptext)
@@ -89,6 +98,29 @@ local function queryproto(self, pname)
 
 	return v
 end
+
+function sproto:request_encode(protoname, tbl)
+	local p = queryproto(self, protoname)
+	return core.encode(p.request,tbl) , p.tag
+end
+
+function sproto:response_encode(protoname, tbl)
+	local p = queryproto(self, protoname)
+	return core.encode(p.response,tbl)
+end
+
+function sproto:request_decode(protoname, ...)
+	local p = queryproto(self, protoname)
+	return core.decode(p.request,...) , p.name
+end
+
+function sproto:response_decode(protoname, ...)
+	local p = queryproto(self, protoname)
+	return core.decode(p.response,...)
+end
+
+sproto.pack = core.pack
+sproto.unpack = core.unpack
 
 local header_tmp = {}