+++ test/testudp.lua
@@ -3,16 +3,15 @@ local socket = require "socket"
 
 local function server()
 	local host
-	host = socket.udp(function(data, sz, from)
-		local str = skynet.tostring(data,sz)	-- skynet.tostring should call only once, because it will free the pointer data
+	host = socket.udp(function(str, from)
 		print("server recv", str, socket.udp_address(from))
 		socket.sendto(host, from, "OK " .. str)
 	end , "127.0.0.1", 8765)	-- bind an address
 end
 
 local function client()
-	local c = socket.udp(function(data, sz, from)
-		print("client recv", skynet.tostring(data,sz), socket.udp_address(from))
+	local c = socket.udp(function(str, from)
+		print("client recv", str, socket.udp_address(from))
 	end)
 	socket.udp_connect(c, "127.0.0.1", 8765)
 	for i=1,20 do